Split file into source map cache
